pocs/_template/requirements.txt
anthropic==0.49.0
python-dotenv==1.0.1
pytest==8.3.5
